requirements.txt
os
opencv-python>=4.7.0.68
numpy
Pillow
easyocr
ultralytics>=8.0.0
streamlit
pymssql
pymongo
gridfs